fix(slam): fix Hesai PTP timestamp truncation in velodyne handler
Hesai XT16 publishes `timestamp` as FLOAT64 (absolute Unix epoch, ~1.7e9 s).
The velodyne_ros::Point struct stored it as `float`, causing PCL to quantize
the value to ~128 s resolution — all per-point offsets collapse to zero.
FAST-LIO2 curvature (expected: relative offset 0–100 ms) received a constant
~1.7e9 value, breaking motion undistortion and causing IKFoM divergence.

- Change `velodyne_ros::Point::time` float→double
- Update POINT_CLOUD_REGISTER_POINT_STRUCT to double
- Compute t0 = points[0].time; convert (time-t0)*1e3 -> ms in both
  feature_enabled and default handler branches

+++ src/fast_lio_ros2/src/preprocess.h
@@ -45,7 +45,7 @@ namespace velodyne_ros {
 struct EIGEN_ALIGN16 Point {
   PCL_ADD_POINT4D;
   float intensity;
-  float time;
+  double time;  // Hesai PTP: absolute Unix epoch [s]; subtracted to relative offset in handler
   uint16_t ring;
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 };
@@ -53,7 +53,7 @@ struct EIGEN_ALIGN16 Point {
 POINT_CLOUD_REGISTER_POINT_STRUCT(
     velodyne_ros::Point,
     (float, x, x)(float, y, y)(float, z, z)(float, intensity, intensity)(
-        float, time, timestamp)(uint16_t, ring, ring))
+        double, time, timestamp)(uint16_t, ring, ring))
 
 namespace ouster_ros {
 struct EIGEN_ALIGN16 Point {
